Parámetros actualizados

+++ 1-introduccion/tema8/Objetos.java
@@ -32,7 +32,7 @@ class Persona {
 	public long getTelefono() {
 		return telefono;
 	}
-	public void setTelefono(int telefono) {
+	public void setTelefono(long telefono) {
 		this.telefono = telefono;
 	}
